Control Updating and AI adding
Updated the controls a bit to smooth out, it still needs fixing though. And added in the start of the AI controls with the ability to patrol, idle and investigate the noise of the player. Will be adding in attacking and other things as time goes on. Also moved the asset pack to third person folder.