More Updates, Floor 3, Still trying hitbox
Struggling with Hitbox still...can't figure it out at all, decided to build floor 3 design. Adding some cool platform to get up, and also added the use of double jump for the player. And added a new blueprint so I can trap each hole as the player falls down it.