Weapon/Ai/Hitbox continuation Part 1
Continued and finally got all hitboxes on my attack animations, it is still a bit mess where both are showing up. And also working on adding damage to the enemy, it doesn't work atm so working on that as well. And added a new trace data to help better trace when I swap weapons, and a new interface to help that. Along with some dropped functions and other things.