Hitbox completed, damage on player and enemy completed, HP Loss on both. Stamina system completed, death animation added
Sooooo a long day I had, this is part 2 for this day. I finished up the hitbox so now it properly fully sphere traces on animation, and only hits 1 time per enemy. Which required changing up the montages, and I added in a widget to show player HP and stamina that fall and regen stamina. I also added in a death animation and gonna start working on that.