New Object
Student Jamile Rendon Torres gave me their object that they were working on, had to fix it up a bit in blender because it didn't import for UE5, and one google search away with the error told me to fix it in blender and it worked!

Thank you again Jamile!