Widgets and Animations
Currently put in more widgets, added in a death screen on player death so they can restart the level or quit. And I tried to make a fade to black screen to go with it using the death animation, but sadly couldn't figure it out maybe will go back to it later. Also started to clean up the enemy animation and now have them swinging on animation just like the player, but will need to fix them being able to hit multiple times and now figured out a bug with the player taking no damage now....Which seems to possibly be with the dodge? Will have to look further.