Ranged AI and Base AI Clean Up
Cleaned up the base ai attack so now it only attacks the player 1 time per trace. And have started to work on the ranged enemy, creating it's behavior tree and getting it's ranged projectile ready. It currently fires but it does no damage to the player even though I coded it in. Will have to keep on testing to get it to work.